Entire JavaScript code for Reviews Project
This is the entire JavaScript code for the Reviews Project where we can click on the Next and Previous Buttons to take us to other reviews. And a Surprise Me button to randomly show a review.

NEW THING LEARNED:
window.AddEventListener("DOMContentLoaded", function() {})

this adds code or things to do when the browser window initially loads the content in the DOM.